+++ tests/src/Rector/PHPUnit/PhpUnitTestAnnotationToAttributeRector/config/configured_rule.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+use DrupalRector\Rector\PHPUnit\PhpUnitTestAnnotationToAttributeRector;
+use DrupalRector\Rector\PHPUnit\ValueObject\PhpUnitTestAnnotationToAttributeConfiguration;
+use DrupalRector\Tests\Rector\Deprecation\DeprecationBase;
+use Rector\Config\RectorConfig;
+
+return static function (RectorConfig $rectorConfig): void {
+    DeprecationBase::addClass(PhpUnitTestAnnotationToAttributeRector::class, $rectorConfig, false, [
+        new PhpUnitTestAnnotationToAttributeConfiguration('11.0.0', '12.0.0', 'group', 'PHPUnit\Framework\Attributes\Group'),
+    ]);
+};